app/requirements.txt
tensorflow==2.20.0
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy

